+++ src/dashboard/helpers/find-possible-long-words.helper.ts
@@ -1,0 +1,11 @@
+import slowa10 from '../../assets/slowa10'
+import slowa11 from '../../assets/slowa11'
+import slowa12 from '../../assets/slowa12'
+import slowa13 from '../../assets/slowa13'
+import slowa14 from '../../assets/slowa14'
+import slowa15 from '../../assets/slowa15'
+
+export const longWordsByLength = [
+  '', '', '', '', '', '', '', '', '', '',
+  slowa10, slowa11, slowa12, slowa13, slowa14, slowa15,
+]