+++ src/dictionary/components/dictionary-buttons/dictionary-buttons.styled.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components/native'
+
+export const DictionarySearchButtonContainer = styled.View`
+  margin-top: 10px;
+  padding-bottom: 10px;
+`