+++ src/dictionary/components/dictionary-buttons/dictionary-buttons.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import { DictionarySearchButtonContainer } from './dictionary-buttons.styled'
+import { CustomButton } from '../../../core/custom-button/custom-button'
+import { SearchButtonIcon } from '../../../dashboard/dashboard.styled'
+
+interface Props {
+  onSearch: () => void;
+}
+
+export const DictionaryButtons = ({ onSearch }: Props) => (
+  <DictionarySearchButtonContainer>
+    <CustomButton onPress={onSearch} withHaptic>
+      <SearchButtonIcon />
+    </CustomButton>
+  </DictionarySearchButtonContainer>
+)